requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
httplib2==0.22.0
idna==3.10
oauth2client==4.1.3
proto-plus==1.26.1
protobuf==6.30.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyDrive==1.3.1
pyparsing==3.2.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
schedule==1.2.2
six==1.17.0
uritemplate==4.1.1
urllib3==2.3.0
